correction of lining skipped in MainMenu

+++ src/business/IEmployeService.java
@@ -2,9 +2,9 @@ package business;
 
 public interface IEmployeService {
 
-	public void AjouterEmploye(String name, String post, Double salary, boolean active);
+	public void AjouterEmploye(String name, String post, int salary, boolean active);
 
-	public void MAJEmploye(int id, String name, String post, Double salary, boolean active);
+	public void MAJEmploye(int id, String name, String post, int salary, boolean active);
 
 	public void SupprimerEmploye(int id);
